backend/pyproject.toml
[project]
name = "cheaters-dilemma-backend"
version = "0.1.0"
description = "Backend for The Cheater's Dilemma multi-agent simulation"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "pydantic>=2.5.0",
  "pydantic-settings>=2.0.0",
  "pyyaml>=6.0.0",
  "python-multipart>=0.0.6",
  "web3>=6.0.0",
  "eth-account>=0.10.0",
  "python-dotenv>=1.2.1",
  "py-solc-x>=2.0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
